+++ common/core/memory_subsystem/page_stats.h
@@ -8,18 +8,18 @@
 class PageStats {
 public:
     PageStats(bool ejecucion, core_id_t core_id);
-    void updatePageStats(int pageAdress, bool evict = false);
+    void updatePageStats(IntPtr pageAdress, bool evict = false);
 
 private:
     std::unordered_map<int, std::bitset<1024>> pageMap; // Asumimos que el tamano de la pagina es 4096 bytes (1024 palabras de 4 bytes)
     std::unordered_map<int, int> pageAccessCount;
-    int totalPages;
-    int minUsage;
-    int maxUsage;
-    int totalAccesses;
-    int totalUse;
+    UInt64 totalPages;
+    UInt64 minUsage;
+    UInt64 maxUsage;
+    UInt64 totalAccesses;
+    UInt64 totalUse;
     bool ejecucion;
-    void evictPage(int pageAddress);
+    void evictPage(IntPtr pageAddress);
 };
 
 #endif // PAGE_STATS_H
